Extract shared fixtures from test classes
