+++ tests/Pipaslot.Mediator.Tests/Mediator_ExecutionExceptionSubtypeTests.cs
@@ -13,7 +13,8 @@ namespace Pipaslot.Mediator.Tests;
 /// <c>Response</c> is populated the same way regardless of subtype - so they don't belong to any single trigger's
 /// own test file. Each trigger's own "throws the correct subtype" assertion lives with that trigger's test
 /// (no-handler in E2E.NoHandlerTests, status-Failed-without-throw in E2E.NoHandlerWithoutErrorTests, missing-result in
-/// E2E.ResultWasTakenFromTheContextTests) - don't duplicate that here.
+/// E2E.ResultWasTakenFromTheContextTests) - don't duplicate that here. Shared middleware fixtures live in
+/// E2E.Fixtures.
 /// </summary>
 public class Mediator_ExecutionExceptionSubtypeTests
 {
@@ -25,8 +26,8 @@ public class Mediator_ExecutionExceptionSubtypeTests
     {
         var sut = trigger switch
         {
-            SubtypeTrigger.MissingResult => Factory.CreateConfiguredMediator(c => c.Use<E2E.ResultWasTakenFromTheContextTests.RemoveResultFromContextMiddleware>()),
-            SubtypeTrigger.UnhandledError => Factory.CreateConfiguredMediator(c => c.Use<E2E.NoHandlerWithoutErrorTests.BlockRequestMiddleware>()),
+            SubtypeTrigger.MissingResult => Factory.CreateConfiguredMediator(c => c.Use<E2E.Fixtures.RemoveResultFromContextMiddleware>()),
+            SubtypeTrigger.UnhandledError => Factory.CreateConfiguredMediator(c => c.Use<E2E.Fixtures.BlockRequestMiddleware>()),
             _ => Factory.CreateConfiguredMediator(),
         };
 
@@ -35,7 +36,7 @@ public class Mediator_ExecutionExceptionSubtypeTests
         {
             if (trigger == SubtypeTrigger.UnhandledError)
             {
-                await sut.ExecuteUnhandled(new E2E.NoHandlerWithoutErrorTests.BlockedRequest());
+                await sut.ExecuteUnhandled(new E2E.Fixtures.BlockedRequest());
             }
             else
             {
@@ -71,13 +72,13 @@ public class Mediator_ExecutionExceptionSubtypeTests
         // Response is built by the MediatorExecutionException base constructor, unchanged by this unit - this test
         // proves that reading Response off the now-thrown MediatorUnhandledErrorException still reflects the real
         // pipeline results exactly as it did off the base type before this change (no regression in what a caught
-        // exception exposes). Reuses NoHandlerAndErrorReturnedTests' fixture rather than a local copy.
-        var sut = Factory.CreateConfiguredMediator(c => c.Use<E2E.NoHandlerAndErrorReturnedTests.AddErrorAndEndMiddleware>());
+        // exception exposes). Reuses the shared AddErrorAndEndMiddleware fixture rather than a local copy.
+        var sut = Factory.CreateConfiguredMediator(c => c.Use<E2E.Fixtures.AddErrorAndEndMiddleware>());
         var action = new SingleHandler.Request(true);
 
         var ex = await Assert.ThrowsAsync<MediatorUnhandledErrorException>(() => sut.ExecuteUnhandled(action));
 
         Assert.False(ex.Response.Success);
-        Assert.Contains(ex.Response.Results.OfType<Notification>(), n => n.Content == E2E.NoHandlerAndErrorReturnedTests.Error);
+        Assert.Contains(ex.Response.Results.OfType<Notification>(), n => n.Content == E2E.Fixtures.AddErrorAndEndMiddleware.Error);
     }
 }
